Finished implementing Device* classes and changing rt classes to use them

+++ include/cl-gl/opengl-init.hpp
@@ -42,7 +42,7 @@ struct GLInfo
 };
 
 GLint init_gl(int argc, char** argv, GLInfo* glinfo, 
-	      const uint32_t* window_size, const std::string title = std::string());
+	      const size_t* window_size, const std::string title = std::string());
 GLuint create_tex_gl(uint32_t width, uint32_t height);
 GLuint create_tex_gl_from_jpeg(uint32_t& width, uint32_t& height, const char* file);
 GLuint create_buf_gl(uint32_t buf_size);
